pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page-analyzer for hexlet"
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
validators = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
